Rename packages and fix context
