+++ dukes-liberty/src/main/java/eu/agilejava/dukes/HealthEndpoint.java
@@ -1,4 +1,4 @@
-package application.rest;
+package eu.agilejava.dukes;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
