+++ dukes-liberty/src/main/webapp/WEB-INF/ibm-web-ext.xml
@@ -4,5 +4,5 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://websphere.ibm.com/xml/ns/javaee http://websphere.ibm.com/xml/ns/javaee/ibm-web-ext_1_0.xsd"
     version="1.0">
-  <context-root uri="/dukesliberty"/>
+  <context-root uri="/dukes-liberty"/>
 </web-ext>
